Update tests for moved code.
