+++ kolibri/utils/tests/test_cli.py
@@ -9,10 +9,8 @@ import os
 import tempfile
 
 import pytest
-from django.db.utils import OperationalError
 from mock import patch
 
-import kolibri
 from kolibri.plugins.utils import autoremove_unavailable_plugins
 from kolibri.utils import cli
 from kolibri.utils import options
@@ -200,78 +198,6 @@ def test_kolibri_listen_port_env(monkeypatch):
         assert "successfully been stopped" in LOG_LOGGER[-1][1]
 
 
-@pytest.mark.django_db
-@patch("kolibri.utils.cli.get_version", return_value="")
-@patch("kolibri.utils.cli.update")
-@patch("kolibri.utils.cli.plugin.callback")
-@patch("kolibri.core.deviceadmin.utils.dbbackup")
-def test_first_run(dbbackup, plugin, update, get_version):
-    """
-    Tests that the first_run() function performs as expected
-    """
-
-    cli.initialize()
-    update.assert_called_once()
-    dbbackup.assert_not_called()
-
-    # Check that it got called for each default plugin
-    from kolibri import plugins
-
-    assert set(plugins.config["INSTALLED_PLUGINS"]) == set(plugins.DEFAULT_PLUGINS)
-
-
-@pytest.mark.django_db
-@patch("kolibri.utils.cli.get_version", return_value="0.0.1")
-@patch("kolibri.utils.cli.update")
-def test_update(update, get_version):
-    """
-    Tests that update() function performs as expected
-    """
-    cli.initialize()
-    update.assert_called_once()
-
-
-@pytest.mark.django_db
-@patch("kolibri.utils.cli.get_version", return_value="0.0.1")
-def test_update_exits_if_running(get_version):
-    """
-    Tests that update() function performs as expected
-    """
-    with patch("kolibri.utils.cli.server.get_status"):
-        try:
-            cli.initialize()
-            pytest.fail("Update did not exit when Kolibri was already running")
-        except SystemExit:
-            pass
-
-
-@pytest.mark.django_db
-def test_version_updated():
-    """
-    Tests our db backup logic: version_updated gets any change, backup gets only non-dev changes
-    """
-    assert cli.version_updated("0.10.0", "0.10.1")
-    assert not cli.version_updated("0.10.0", "0.10.0")
-    assert not cli.should_back_up("0.10.0-dev0", "")
-    assert not cli.should_back_up("0.10.0-dev0", "0.10.0")
-    assert not cli.should_back_up("0.10.0", "0.10.0-dev0")
-    assert not cli.should_back_up("0.10.0-dev0", "0.10.0-dev0")
-
-
-@pytest.mark.django_db
-@patch("kolibri.utils.cli.get_version", return_value=kolibri.__version__)
-@patch("kolibri.utils.cli.update")
-@patch("kolibri.core.deviceadmin.utils.dbbackup")
-def test_update_no_version_change(dbbackup, update, get_version):
-    """
-    Tests that when the version doesn't change, we are not doing things we
-    shouldn't
-    """
-    cli.initialize()
-    update.assert_not_called()
-    dbbackup.assert_not_called()
-
-
 def test_cli_usage():
     # Test the -h
     with pytest.raises(SystemExit) as excinfo:
@@ -305,18 +231,3 @@ def test_list_plugins_disabled(echo_mock, plugins):
             echo_mock.call_args_list,
         )
     )
-
-
-@patch("kolibri.utils.cli._migrate_databases")
-@patch("kolibri.utils.cli.version_updated")
-def test_migrate_if_unmigrated(version_updated, _migrate_databases):
-    # No matter what, ensure that version_updated returns False
-    version_updated.return_value = False
-    from morango.models import InstanceIDModel
-
-    with patch.object(
-        InstanceIDModel, "get_or_create_current_instance"
-    ) as get_or_create_current_instance:
-        get_or_create_current_instance.side_effect = OperationalError("Test")
-        cli.initialize()
-        _migrate_databases.assert_called_once()
